Refactored QRCodeFragment-still need to resolve conflicts in tests
