+++ app/src/main/java/com/cmput301w23t09/qrhunter/scanqr/ScannerController.java
@@ -6,6 +6,7 @@ import android.util.Log;
 import androidx.camera.core.ImageProxy;
 import com.cmput301w23t09.qrhunter.BaseFragment;
 import com.cmput301w23t09.qrhunter.player.Player;
+import com.cmput301w23t09.qrhunter.qrcode.AddQRCodeFragment;
 import com.cmput301w23t09.qrhunter.qrcode.QRCode;
 import com.cmput301w23t09.qrhunter.qrcode.QRCodeFragment;
 import com.cmput301w23t09.qrhunter.scanqr.camera.CameraController;
@@ -83,7 +84,7 @@ public class ScannerController {
                         pastHash = currentHash;
                         if (qrCodeFragment != null) qrCodeFragment.dismissNow();
                         QRCode qrCode = new QRCode(pastHash);
-                        qrCodeFragment = QRCodeFragment.newInstance(qrCode, activePlayer);
+                        qrCodeFragment = AddQRCodeFragment.newInstance(qrCode, activePlayer);
                         fragment.getGameController().setPopup(qrCodeFragment);
                       }
                     }
